cleanup, move repository source to  ipc project
