+++ src/RepoM.App/Converters/UtcToHumanizedLocalDateTimeConverter.cs
@@ -7,21 +7,21 @@ using RepoM.Api.Common;
 
 public class UtcToHumanizedLocalDateTimeConverter : IValueConverter
 {
+    private readonly IHumanizer _humanizer;
+
     public UtcToHumanizedLocalDateTimeConverter()
     {
-        Humanizer = new HardcodededMiniHumanizer();
+        _humanizer = new HardcodededMiniHumanizer();
     }
 
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {
         DateTime date = DateTime.SpecifyKind(DateTime.Parse(value.ToString() ?? string.Empty), DateTimeKind.Utc).ToLocalTime();
-        return Humanizer.HumanizeTimestamp(date);
+        return _humanizer.HumanizeTimestamp(date);
     }
 
     public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
     {
         throw new NotImplementedException();
     }
-
-    public IHumanizer Humanizer { get; }
 }
